requirements.txt
python-dotenv==1.0.1
python-telegram-bot==21.3
openai==1.40.0
pdfplumber==0.11.4
python-docx==1.1.2
